pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "vyvotts"
version = "0.1.0"
description = "LLM-Based Text-to-Speech Training and Inference Framework"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "transformers",
    "datasets>=3.4",
    "accelerate",
    "snac @ git+https://github.com/kadirnar/fast-snac.git",
    "soundfile",
    "librosa",
    "torchaudio",
]

[project.optional-dependencies]
vllm = ["vllm>=0.18"]
sglang = ["sglang[all]", "sgl-kernel"]
train = ["unsloth", "wandb", "trl", "kernels", "liger_kernel"]

[tool.setuptools.packages.find]
include = ["vyvotts*"]
